Outsourced preferences into its own service
