+++ src/app/internal-services/preferences.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from "@angular/core/testing";
+
+import { PreferencesService } from "./preferences.service";
+
+describe("PreferencesService", () => {
+    let service: PreferencesService;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({});
+        service = TestBed.inject(PreferencesService);
+    });
+
+    it("should be created", () => {
+        expect(service).toBeTruthy();
+    });
+});
